refactor(modularization): Phase 4 Step 4-S2d — model_platform DB split (LLMProvider + LLMProviderModel + APIType → aperag/domains/model_platform/db/models.py) (#25)
Fourth and final sub-step of Phase 4 Step 4-S2.

**New module** ``aperag/domains/model_platform/db/models.py``:
- ``LLMProvider`` + ``LLMProviderModel`` class bodies moved
  verbatim (all 5 helper methods on ``LLMProviderModel`` kept:
  ``get_provider`` / ``set_provider`` / ``has_tag`` / ``add_tag``
  / ``remove_tag`` / ``get_tags``).
- ``APIType`` enum moved alongside (only ``LLMProviderModel.api``
  uses it).
- Local ``_random_id`` / ``_enum_column`` helpers.
- Docstring reiterates that ``aperag.llm.*`` is shared
  infrastructure and does NOT move into this domain (Phase 4
  canonical msg=d47fa490 Section 7).
- Docstring also notes ``ModelServiceProvider`` is a separate
  legacy table that remains in ``aperag.db.models`` — Phase 6
  cleanup decides its fate.

**aperag/db/models.py**:
- Class bodies replaced with breadcrumbs.
- Re-export block extended with the three model_platform symbols.

**Phase 4 4-S2 complete**: the re-export shim now covers all 13
Phase 4 DB symbols (3 identity + 4 governance + 3 marketplace + 3
model_platform), extending Phase 3's 15 to a running total of 28.

**Verification**:
- ``llm_provider`` / ``llm_provider_models`` tables still on
  ``Base.metadata``.
- All 13 Phase 4 DB symbols importable via
  ``aperag.db.models``.
- Full unit suite green: 671 passed / 29 skipped.
- ``ruff check`` + ``ruff format --check`` clean.
